lowered capture rate from 1 frame / second to 1 frame / 2 seconds, disabled swap, hopefully i/o should be quick enough to handle capture rate

+++ src/FCamera.cpp
@@ -134,7 +134,7 @@ void FCamera::expose()
     autoWhiteBalance(&shot, frame);
 }
 
-long FCamera::getMaximumExposureTime()
+double FCamera::getMaximumExposureTime()
 {
     if(day)
         return(MAXIMUM_DAY_EXPOSURE);
